requirements.txt
# Core dependencies with pinned versions for reproducibility
# Python 3.10+ required

# Deep Learning & RL
torch==2.1.2
transformers==4.36.2
accelerate==0.25.0
peft==0.7.1
datasets==2.16.1
stable-baselines3==2.2.1
gymnasium==0.29.1

# Data & Finance
numpy==1.24.3
pandas==2.0.3
yfinance==0.2.33
pandas-datareader==0.10.0
fredapi==0.5.1
sec-edgar-downloader==5.0.3
ta==0.11.0  # Technical analysis library

# Backtesting & Trading
backtrader==1.9.78.123
bt==0.2.9
zipline-reloaded==3.0.3
pyfolio-reloaded==0.9.5

# Visualization
matplotlib==3.7.3
seaborn==0.13.0
plotly==5.18.0
graphviz==0.20.1

# Statistics & ML
scikit-learn==1.3.2
scipy==1.11.4
statsmodels==0.14.1

# NLP & Text Processing
nltk==3.8.1
spacy==3.7.2
sentencepiece==0.1.99

# LLM APIs (optional, for cloud models)
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
wandb==0.16.2
tensorboard==2.15.1
pyyaml==6.0.1
click==8.1.7
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# API & Web
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1

# Experiment tracking
mlflow==2.9.2
shap==0.45.0
alibi==0.9.6
litellm==1.52.3
alpaca-py==0.24.0
ib_insync==0.9.86
pydantic==2.6.1
rich==13.7.1
